move recipients and groups to its own component update TR
